Adding initial loadbalancer controller with Openstack Lbaas and nginx backends
This controller is designed to easily integrate and create different load balancing backends.
From software, hardware to cloud loadbalancers.

For software loadbalancer, keepalived is used to manage VIPs so that software loadbalancers
can bind to them.

The software loadbalancers run in nodes with keepalived. They communicate with the loadbalancer controller
via configmaps.

- Added backend for provisioning Openstack Lbaas
- Nginxs added as loadbalancer-daemon backend
- Added a keepalived controller to manage VIPs for nginx backend
- Updated Dockerfile to include keepalived and its dependency packages
- Added example files
- Added README